+++ app/src/main/java/com/dskja/betterstreamflix/fragments/settings/SettingsTvFragment.kt
@@ -677,6 +677,22 @@ class SettingsTvFragment : LeanbackPreferenceFragmentCompat() {
             }
         }
 
+        findPreference<SwitchPreference>("DOWNLOAD_SMART_ENABLED")?.apply {
+            isChecked = UserPreferences.downloadSmartEnabled
+            setOnPreferenceChangeListener { _, newValue ->
+                UserPreferences.downloadSmartEnabled = newValue as Boolean
+                true
+            }
+        }
+
+        findPreference<SwitchPreference>("DOWNLOAD_AUTO_DELETE_WATCHED")?.apply {
+            isChecked = UserPreferences.downloadAutoDeleteWatched
+            setOnPreferenceChangeListener { _, newValue ->
+                UserPreferences.downloadAutoDeleteWatched = newValue as Boolean
+                true
+            }
+        }
+
         findPreference<Preference>("DOWNLOAD_STORAGE_USED")?.summary =
             DownloadStorage.formatBytes(DownloadStorage.usedBytes(requireContext()))
 
@@ -2153,6 +2169,8 @@ class SettingsTvFragment : LeanbackPreferenceFragmentCompat() {
         findPreference<SwitchPreference>("FORCE_EXTRA_BUFFERING")?.isChecked = UserPreferences.forceExtraBuffering
         findPreference<SwitchPreference>("SERVER_AUTO_SUBTITLES_DISABLED")?.isChecked = UserPreferences.serverAutoSubtitlesDisabled
         findPreference<SwitchPreference>("DOWNLOAD_WIFI_ONLY")?.isChecked = UserPreferences.downloadWifiOnly
+        findPreference<SwitchPreference>("DOWNLOAD_SMART_ENABLED")?.isChecked = UserPreferences.downloadSmartEnabled
+        findPreference<SwitchPreference>("DOWNLOAD_AUTO_DELETE_WATCHED")?.isChecked = UserPreferences.downloadAutoDeleteWatched
         findPreference<SwitchPreference>("DOWNLOAD_NOTIFY_COMPLETE")?.isChecked = UserPreferences.downloadNotifyComplete
         findPreference<SwitchPreference>("DOWNLOAD_FILTER_CURRENT_PROVIDER")?.isChecked =
             UserPreferences.downloadFilterCurrentProvider